Unify fill-advert-slots and fill-external-slots
